+++ packages/infra/src/api/routing/makeRequestHandler.ts
@@ -12,7 +12,7 @@ import { Effect, FiberRef, Option, S } from "effect-app"
 import type { HttpServerError } from "effect-app/http"
 import { HttpBody, HttpRouter, HttpServerRequest, HttpServerResponse } from "effect-app/http"
 import { NonEmptyString255 } from "effect-app/schema"
-import type { REST, Schema, StructFields } from "effect-app/schema"
+import type { REST, Struct } from "effect-app/schema"
 import { updateRequestContext } from "../setupRequest.js"
 import { makeRequestParsers, parseRequestParams } from "./base.js"
 import type { RequestHandler, RequestHandlerBase } from "./base.js"
@@ -24,13 +24,13 @@ export const RequestSettings = FiberRef.unsafeMake({
 export type Middleware<
   R,
   M,
-  PathA extends StructFields,
-  CookieA extends StructFields,
-  QueryA extends StructFields,
-  BodyA extends StructFields,
-  HeaderA extends StructFields,
+  PathA extends Struct.Fields,
+  CookieA extends Struct.Fields,
+  QueryA extends Struct.Fields,
+  BodyA extends Struct.Fields,
+  HeaderA extends Struct.Fields,
   ReqA extends PathA & QueryA & BodyA,
-  ResA extends StructFields,
+  ResA extends Struct.Fields,
   ResE,
   MiddlewareE,
   PPath extends `/${string}`,
@@ -79,13 +79,13 @@ export type Middleware<
 export function makeRequestHandler<
   R,
   M,
-  PathA extends StructFields,
-  CookieA extends StructFields,
-  QueryA extends StructFields,
-  BodyA extends StructFields,
-  HeaderA extends StructFields,
+  PathA extends Struct.Fields,
+  CookieA extends Struct.Fields,
+  QueryA extends Struct.Fields,
+  BodyA extends Struct.Fields,
+  HeaderA extends Struct.Fields,
   ReqA extends PathA & QueryA & BodyA,
-  ResA extends StructFields,
+  ResA extends Struct.Fields,
   ResE,
   MiddlewareE,
   R2,
@@ -128,15 +128,15 @@ export function makeRequestHandler<
   const { Request, Response, h: handle } = handler
 
   const response: REST.ReqRes<any, any, any> = Response ? Response : S.void
-  const resp = response as typeof response & { struct?: Schema<any, any, any> }
+  const resp = response as typeof response & { fields?: S.Struct.Fields }
   // TODO: consider if the alternative of using the struct schema is perhaps just better.
-  const encoder = "struct" in resp && resp.struct
-    ? S.encode(handler.rt === "raw" ? S.from(resp.struct) : resp.struct)
+  const encoder = "fields" in resp && resp.fields
+    ? S.encode(handler.rt === "raw" ? S.encodedSchema(S.struct(resp.fields)) : S.struct(resp.fields))
     // ? (i: any) => {
     //   if (i instanceof (response as any)) return S.encodeSync(response)(i)
     //   else return S.encodeSync(response)(new (response as any)(i))
     // }
-    : S.encode(handler.rt === "raw" ? S.from(resp) : resp)
+    : S.encode(handler.rt === "raw" ? S.encodedSchema(resp) : resp)
   // const encodeResponse = adaptResponse
   //   ? (req: ReqA) => Encoder.for(adaptResponse(req))
   //   : () => encoder