+++ packages/schema/src/ext.ts
@@ -1,57 +1,78 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable @typescript-eslint/no-unsafe-return */
-import { Effect } from "@effect-app/core"
-import { extendM, typedKeysOf } from "@effect-app/core/utils"
-import type { Schema, StructFields } from "@effect/schema/Schema"
+import type { NonEmptyReadonlyArray } from "@effect-app/core"
+import { Effect, pipe } from "@effect-app/core"
+import { extendM } from "@effect-app/core/utils"
+import type { Schema } from "@effect/schema/Schema"
 import * as S from "@effect/schema/Schema"
 import { flow } from "effect"
 import type { AST, ParseResult } from "./index.js"
 
-export const Date = Object.assign(S.Date, { withDefault: S.withDefaultConstructor(S.Date, () => new global.Date()) })
-export const boolean = Object.assign(S.boolean, { withDefault: S.withDefaultConstructor(S.boolean, () => false) })
-export const number = Object.assign(S.number, { withDefault: S.withDefaultConstructor(S.number, () => 0) })
+export const Date = Object.assign(S.Date, {
+  withDefault: S.propertySignature(S.Date, { default: () => new global.Date() })
+})
+export const boolean = Object.assign(S.boolean, {
+  withDefault: S.propertySignature(S.boolean, { default: () => false })
+})
+export const number = Object.assign(S.number, { withDefault: S.propertySignature(S.number, { default: () => 0 }) })
 
 /**
  * Like the default Schema `struct` but with batching enabled by default
  */
-export const struct = flow(
-  S.struct,
-  S.batching(true)
-)
+export function struct<Fields extends S.Struct.Fields, const Records extends S.IndexSignature.NonEmptyRecords>(
+  fields: Fields,
+  ...records: Records
+): S.typeLiteral<Fields, Records>
+export function struct<Fields extends S.Struct.Fields>(fields: Fields): S.struct<Fields>
+export function struct<Fields extends S.Struct.Fields, const Records extends S.IndexSignature.Records>(
+  fields: Fields,
+  ...records: Records
+): S.typeLiteral<Fields, Records> {
+  return S.struct(fields, ...records as any).pipe(S.batching(true))
+}
 
 /**
  * Like the default Schema `tuple` but with batching enabled by default
  */
-export const tuple = flow(
-  S.tuple,
-  S.batching(true)
-)
+export function tuple<
+  const Elements extends S.TupleType.Elements,
+  Rest extends NonEmptyReadonlyArray<Schema.Any>
+>(elements: Elements, ...rest: Rest): S.tupleType<Elements, Rest>
+export function tuple<Elements extends S.TupleType.Elements>(...elements: Elements): S.tuple<Elements>
+export function tuple(...args: ReadonlyArray<any>): any {
+  return S.tuple(...args).pipe(S.batching(true))
+}
 
 /**
  * Like the default Schema `nonEmptyArray` but with batching enabled by default
  */
-export const nonEmptyArray = flow(
-  S.nonEmptyArray,
-  S.batching(true)
-)
+export function nonEmptyArray<Value extends Schema.Any>(value: Value): S.nonEmptyArray<Value> {
+  return pipe(
+    S.nonEmptyArray(value),
+    S.batching(true)
+  )
+}
 
 /**
  * Like the default Schema `array` but with `withDefault` and batching enabled by default
  */
-export const array = flow(
-  S.array,
-  S.batching(true),
-  (s) => Object.assign(s, { withDefault: S.withDefaultConstructor(s, () => []) })
-)
+export function array<Value extends Schema.Any>(value: Value) {
+  return pipe(
+    S.array(value),
+    S.batching(true),
+    (s) => Object.assign(s, { withDefault: S.propertySignature(s, { default: () => [] as any }) })
+  )
+}
 
 /**
  * Like the default Schema `readonlySet` but with `withDefault` and batching enabled by default
  */
-export const readonlySet = flow(
-  S.readonlySet,
-  S.batching(true),
-  (s) => Object.assign(s, { withDefault: S.withDefaultConstructor(s, () => new Set()) })
-)
+export const readonlySet = <Value extends Schema.Any>(value: Value) =>
+  pipe(
+    S.readonlySet(value),
+    S.batching(true),
+    (s) => Object.assign(s, { withDefault: S.propertySignature(s, { default: () => new Set<S.Schema.Type<Value>>() }) })
+  )
 
 /**
  * Like the default Schema `readonlyMap` but with `withDefault` and batching enabled by default
@@ -59,7 +80,7 @@ export const readonlySet = flow(
 export const readonlyMap = flow(
   S.readonlyMap,
   S.batching(true),
-  (s) => Object.assign(s, { withDefault: S.withDefaultConstructor(s, () => new Map()) })
+  (s) => Object.assign(s, { withDefault: S.propertySignature(s, { default: () => new Map() }) })
 )
 
 /**
@@ -67,27 +88,27 @@ export const readonlyMap = flow(
  */
 export const nullable = flow(
   S.nullable,
-  (s) => Object.assign(s, { withDefault: S.withDefaultConstructor(s, () => null) })
+  (s) => Object.assign(s, { withDefault: S.propertySignature(s, { default: () => null }) })
 )
 
 export const defaultDate = <S extends Schema<Date, any, any>>(s: S) =>
-  S.withDefaultConstructor(s, () => new global.Date() as any) // TODO
+  S.propertySignature(s, { default: () => new global.Date() as any }) // }TODO
 
 export const defaultBool = <S extends Schema<boolean, any, any>>(s: S) =>
-  S.withDefaultConstructor(s, () => false as any) // TODO
+  S.propertySignature(s, { default: () => false as any }) // }TODO
 
 export const defaultNullable = <S extends Schema<any, any, any>, From, To>(
   s: S & Schema<To | null, From, Schema.Context<S>>
-) => S.withDefaultConstructor(s, () => null as any) // TODO
+) => S.propertySignature(s, { default: () => null as any }) // }TODO
 
 export const defaultArray = <S extends Schema<ReadonlyArray<any>, any, any>>(s: S) =>
-  S.withDefaultConstructor(s, () => [] as any) // TODO
+  S.propertySignature(s, { default: () => [] as any }) // }TODO
 
 export const defaultMap = <S extends Schema<ReadonlyMap<any, any>, any, any>>(s: S) =>
-  S.withDefaultConstructor(s, () => new Map() as any) // TODO
+  S.propertySignature(s, { default: () => new Map() as any }) // }TODO
 
 export const defaultSet = <S extends Schema<ReadonlySet<any>, any, any>>(s: S) =>
-  S.withDefaultConstructor(s, () => new Set() as any) // TODO
+  S.propertySignature(s, { default: () => new Set() as any }) // }TODO
 
 /**
  * @tsplus getter effect/schema/Schema withDefaults
@@ -108,9 +129,9 @@ export const literal = <Literals extends ReadonlyArray<AST.LiteralValue>>(
 ) => Object.assign(S.literal(...literals) as Schema<Literals[number]>, { literals })
 
 export type WithDefaults<Self extends S.Schema<any, any, never>> = (
-  i: S.Schema.From<Self>,
+  i: S.Schema.Encoded<Self>,
   options?: AST.ParseOptions
-) => S.Schema.To<Self>
+) => S.Schema.Type<Self>
 
 // type GetKeys<U> = U extends Record<infer K, any> ? K : never
 // type UnionToIntersection2<U extends object> = {
@@ -128,56 +149,56 @@ export type WithDefaults<Self extends S.Schema<any, any, never>> = (
 
 export const inputDate = extendM(
   S.union(S.ValidDateFromSelf, S.Date),
-  (s) => ({ withDefault: S.withDefaultConstructor(s, () => new global.Date()) })
+  (s) => ({ withDefault: S.propertySignature(s, { default: () => new global.Date() }) })
 )
 
 export interface UnionBrand {}
 
-export function makeOptional<NER extends StructFields>(
-  t: NER // TODO: enforce non empty
-): {
-  [K in keyof NER]: S.PropertySignature<
-    Schema.From<NER[K]> | undefined,
-    true,
-    Schema.To<NER[K]> | undefined,
-    true,
-    Schema.Context<NER[K]>
-  >
-} {
-  return typedKeysOf(t).reduce((prev, cur) => {
-    prev[cur] = S.optional(t[cur] as any)
-    return prev
-  }, {} as any)
-}
+// export function makeOptional<NER extends Struct.Fields>(
+//   t: NER // TODO: enforce non empty
+// ): {
+//   [K in keyof NER]: S.PropertySignature<
+//     Schema.Encoded<NER[K]> | undefined,
+//     true,
+//     Schema.Type<NER[K]> | undefined,
+//     true,
+//     Schema.Context<NER[K]>
+//   >
+// } {
+//   return typedKeysOf(t).reduce((prev, cur) => {
+//     prev[cur] = S.optional(t[cur] as any)
+//     return prev
+//   }, {} as any)
+// }
 
-export function makeExactOptional<NER extends StructFields>(
-  t: NER // TODO: enforce non empty
-): {
-  [K in keyof NER]: S.PropertySignature<
-    Schema.From<NER[K]>,
-    true,
-    Schema.To<NER[K]>,
-    true,
-    Schema.Context<NER[K]>
-  >
-} {
-  return typedKeysOf(t).reduce((prev, cur) => {
-    prev[cur] = S.optional(t[cur] as any, { exact: true })
-    return prev
-  }, {} as any)
-}
+// export function makeExactOptional<NER extends Struct.Fields>(
+//   t: NER // TODO: enforce non empty
+// ): {
+//   [K in keyof NER]: S.PropertySignature<
+//     Schema.Encoded<NER[K]>,
+//     true,
+//     Schema.Type<NER[K]>,
+//     true,
+//     Schema.Context<NER[K]>
+//   >
+// } {
+//   return typedKeysOf(t).reduce((prev, cur) => {
+//     prev[cur] = S.optional(t[cur] as any, { exact: true })
+//     return prev
+//   }, {} as any)
+// }
 
 /** A version of transform which is only a one way mapping of From->To */
-export const transformTo = <FromA, FromI, FromR, ToA, ToI, ToR>(
-  from: Schema<FromA, FromI, FromR>,
-  to: Schema<ToA, ToI, ToR>,
+export const transformTo = <To extends Schema.Any, From extends Schema.Any>(
+  from: From,
+  to: To,
   decode: (
-    fromA: FromA,
+    fromA: Schema.Type<From>,
     options: AST.ParseOptions,
-    ast: AST.Transform
-  ) => ToI
+    ast: AST.Transformation
+  ) => Schema.Encoded<To>
 ) =>
-  S.transformOrFail<FromA, FromI, FromR, ToA, ToI, ToR, never, never>(
+  S.transformOrFail<To, From, never, never>(
     from,
     to,
     (...args) => Effect.sync(() => decode(...args)),
@@ -185,13 +206,12 @@ export const transformTo = <FromA, FromI, FromR, ToA, ToI, ToR>(
   )
 
 /** A version of transformOrFail which is only a one way mapping of From->To */
-export const transformToOrFail = <FromA, FromI, FromR, ToA, ToI, ToR, R3>(
-  from: Schema<FromA, FromI, FromR>,
-  to: Schema<ToA, ToI, ToR>,
+export const transformToOrFail = <To extends Schema.Any, From extends Schema.Any, RD>(
+  from: From,
+  to: To,
   decode: (
-    fromA: FromA,
+    fromA: Schema.Type<From>,
     options: AST.ParseOptions,
-    ast: AST.Transform
-  ) => Effect.Effect<ToI, ParseResult.ParseIssue, R3>
-) =>
-  S.transformOrFail<FromA, FromI, FromR, ToA, ToI, ToR, R3, never>(from, to, decode, () => Effect.die("one way schema"))
+    ast: AST.Transformation
+  ) => Effect.Effect<Schema.Encoded<To>, ParseResult.ParseIssue, RD>
+) => S.transformOrFail<To, From, RD, never>(from, to, decode, () => Effect.die("one way schema"))